Centralize AppKey and publication synchronization logic Standardized configuration management across complex multi-model elements by migrating manual status request logic into a centralized model-layer interface. - Implemented `request_status()` in Light CTL, HSL, Lightness, and Sensor server models. - Implemented model-specific control methods (`request_lightness`, `request_hsl`, etc.) in client models. - Refactored RGB, CWWW, and Sensor elements to utilize the new centralized synchronization APIs, ensuring all child models pull AppKey and publication addresses from the parent element context. - Encapsulated transmission parameter retrieval (`app_id`, `pub_addr`) within model classes to prevent network configuration mismatches in multi-model elements. - Fixed preprocessor scoping (#if CONFIG_...) in model source files to ensure correct class encapsulation and resolve compilation failures across various product profiles. - Resolved syntax errors and missing braces in `meshx_model_ctl.cpp` and `meshx_model_lightness.cpp` resulting from refactoring. - Corrected state pointer casting in `meshx_model_sensor.cpp` to use the appropriate `meshx_sensor_srv_status_t` type. - Cleaned up redundant synchronization loops at the element layer, reducing boilerplate and improving maintainability. This change ensures that all elements correctly report their state using the provisioned AppKey and publication address immediately after a system stack ready or fresh boot event.

+++ main/component/meshx/ble_mesh/common/inc/meshx_fwd_decl.hpp
@@ -74,7 +74,7 @@ using meshxElementType_t = enum meshxElementType;
  * @brief Common element context structure for base handling.
  */
 typedef struct {
-    uint8_t  app_id;
+    uint16_t app_id;
     uint16_t pub_addr;
 } meshx_element_common_ctx_t;
